# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cab Dispatch API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/trips": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Create Trip",
        "operationId": "create_trip_v1_trips_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "List Trips",
        "operationId": "list_trips_v1_trips_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "vehicle_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vehicle Id"
            }
          },
          {
            "name": "driver_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Driver Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/trips/sync": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Sync Trips",
        "operationId": "sync_trips_v1_trips_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/TripSyncItem"
                },
                "type": "array",
                "title": "Items"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/trips/{trip_id}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Trip",
        "operationId": "get_trip_v1_trips__trip_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "trips"
        ],
        "summary": "Update Trip",
        "operationId": "update_trip_v1_trips__trip_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "trips"
        ],
        "summary": "Delete Trip",
        "operationId": "delete_trip_v1_trips__trip_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/trips/{trip_id}/tick": {
      "patch": {
        "tags": [
          "trips"
        ],
        "summary": "Tick Trip",
        "operationId": "tick_trip_v1_trips__trip_id__tick_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripTickRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/trips/{trip_id}/close": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Close Trip Endpoint",
        "operationId": "close_trip_endpoint_v1_trips__trip_id__close_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCloseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TripCreate": {
        "properties": {
          "client_uuid": {
            "type": "string",
            "title": "Client Uuid",
            "description": "Offline idempotency key, unique per tenant"
          },
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "shift_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Id"
          },
          "tariff_id": {
            "type": "string",
            "title": "Tariff Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "rank_hail",
              "booked",
              "airport_fixed",
              "multi_hire"
            ],
            "title": "Type"
          },
          "start_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start At",
            "description": "Defaults to now() if omitted"
          },
          "start_lat": {
            "type": "number",
            "title": "Start Lat"
          },
          "start_lng": {
            "type": "number",
            "title": "Start Lng"
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "cash",
              "card"
            ],
            "title": "Payment Method",
            "default": "cash"
          },
          "time_class": {
            "type": "string",
            "enum": [
              "day",
              "night",
              "holiday"
            ],
            "title": "Time Class",
            "default": "day"
          },
          "is_peak": {
            "type": "boolean",
            "title": "Is Peak",
            "default": false
          },
          "maxi": {
            "type": "boolean",
            "title": "Maxi",
            "default": false
          },
          "tolls": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Tolls",
            "default": "0"
          },
          "extras": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Extras",
            "default": "0"
          },
          "gps_trace_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gps Trace Ref"
          }
        },
        "type": "object",
        "required": [
          "client_uuid",
          "vehicle_id",
          "driver_id",
          "tariff_id",
          "type",
          "start_lat",
          "start_lng"
        ],
        "title": "TripCreate"
      },
      "TripListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TripRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "TripListResponse"
      },
      "TripRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "client_uuid": {
            "type": "string",
            "title": "Client Uuid"
          },
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "shift_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Id"
          },
          "tariff_id": {
            "type": "string",
            "title": "Tariff Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "time_class": {
            "type": "string",
            "title": "Time Class"
          },
          "is_peak": {
            "type": "boolean",
            "title": "Is Peak"
          },
          "maxi": {
            "type": "boolean",
            "title": "Maxi"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End At"
          },
          "start_lat": {
            "type": "number",
            "title": "Start Lat"
          },
          "start_lng": {
            "type": "number",
            "title": "Start Lng"
          },
          "end_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Lat"
          },
          "end_lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Lng"
          },
          "distance_m": {
            "type": "integer",
            "title": "Distance M"
          },
          "moving_s": {
            "type": "integer",
            "title": "Moving S"
          },
          "waiting_s": {
            "type": "integer",
            "title": "Waiting S"
          },
          "flag_fall": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Flag Fall"
          },
          "dist_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Dist Amount"
          },
          "wait_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Wait Amount"
          },
          "peak_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Peak Amount"
          },
          "tolls": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Tolls"
          },
          "psl": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Psl"
          },
          "extras": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Extras"
          },
          "subtotal": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Subtotal"
          },
          "surcharge": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Surcharge"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total"
          },
          "gst_component": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Gst Component"
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method"
          },
          "gps_trace_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gps Trace Ref"
          },
          "max_fare_check_passed": {
            "type": "boolean",
            "title": "Max Fare Check Passed"
          },
          "variance_pct": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variance Pct"
          },
          "receipt_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Ref"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "client_uuid",
          "vehicle_id",
          "driver_id",
          "shift_id",
          "tariff_id",
          "type",
          "status",
          "time_class",
          "is_peak",
          "maxi",
          "start_at",
          "end_at",
          "start_lat",
          "start_lng",
          "end_lat",
          "end_lng",
          "distance_m",
          "moving_s",
          "waiting_s",
          "flag_fall",
          "dist_amount",
          "wait_amount",
          "peak_amount",
          "tolls",
          "psl",
          "extras",
          "subtotal",
          "surcharge",
          "total",
          "gst_component",
          "payment_method",
          "gps_trace_ref",
          "max_fare_check_passed",
          "variance_pct",
          "receipt_ref",
          "created_at",
          "updated_at"
        ],
        "title": "TripRead"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}